Cache auth responses for better performance
The various auth responses have a window of time during which they can
be reused. The most useful use case for using this window of time is to
get the SAML assertion for logging into an AWS console and then, in
other browser container tabs, log into the aws console of other accounts
using the same SAML assertion (and even the AWS CLI).

By using the cached responses we only need to send network traffic for
the first auth interaction and then used the cached response if all
other requests happen within the validity window of time, which speeds
things up considerably.